Mars Rover: Basic "stay within the borders"
We don't rotate yet because we still need to convert rotation angles to rotation duration (or whatever we're using). The DSL should use angles (it's easier) and we also need to take the speed into consideration when calculating this (unles we rotate the wheel based on an angle, and not duration).